no more \lstinline inside \subsection
This was causing extremely difficult-to-trace latex compilation errors
that were triggered by certain pasted python output inside lstlisting
environments, giving "! Extra }, or forgotten $." error messages. The
following stackexchange answer pointed me to the fact that it was
actually not a problem with the block of code in lstlisting, just the
error from \subsection{\lstinline{...}} gets triggered there sometimes
http://tex.stackexchange.com/questions/229113/another-extra-or-forgotten-endgroup
Geez.